[FE][REFACTOR] 타입 스크립트 단언문 및 RTK Query API 리팩토링
