+++ client/src/components/postDetailP/CommentInput.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 import { useParams } from 'react-router-dom';
 import styled from 'styled-components';
-import { commentsApi } from '../../api/api';
+import { commentsApi } from '../../api/commentApi';
 import { useAppDispatch } from '../../hooks';
 import { addCommentEdit, setComment } from '../../slices/commentSlice';
 
@@ -12,8 +12,8 @@ const CommentInput: React.FC = () => {
   const postId = params.postId;
   const query = commentsApi.useGetCommentQuery({ postId });
   // 댓글 추가 mutation
-  const mutation = commentsApi.useSetCommentMutation();
-  const setComments = mutation[0];
+  const commetMutation = commentsApi.useSetCommentMutation();
+  const [setComments] = commetMutation;
 
   // 댓글 추가
   const addCommentHandler = async () => {
