+++ client/src/store/store.tsx
@@ -9,17 +9,15 @@ import headerSlice from '../slices/headerSlice';
 import mainSlice from '../slices/mainSlice';
 import mypageSlice from '../slices/mypageSlice';
 import nicknameSlice from '../slices/nicknameSlice';
-import {
-  postsApi,
-  commentsApi,
-  repliesApi,
-  recomendedPostsApi,
-  reportApi,
-} from '../api/api';
+import { repliesApi } from '../api/replyApi';
 import surveySlice from '../slices/surveySlice';
+import { postsApi, recomendedPostsApi } from '../api/postApi';
+import { commentsApi } from '../api/commentApi';
+import { reportApi } from '../api/reportApi';
 
 const store = configureStore({
   reducer: {
+    // general Reducer
     postInput: postInputSlice.reducer,
     post: postSlice.reducer,
     comment: commentSlice.reducer,
@@ -30,6 +28,8 @@ const store = configureStore({
     mypage: mypageSlice.reducer,
     survey: surveySlice.reducer,
     nickname: nicknameSlice.reducer,
+
+    // api Reducer
     [postsApi.reducerPath]: postsApi.reducer,
     [commentsApi.reducerPath]: commentsApi.reducer,
     [repliesApi.reducerPath]: repliesApi.reducer,
